backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
email-validator
python-multipart
python-jose
passlib[bcrypt]
redis
pymongo
alembic
pytest
httpx
